Fix MetadataMatch.matrix() IndexError when no spectra match
In equal_match mode, empty match lists were converted to float64 arrays, which
cannot index a numpy array, raising IndexError whenever zero pairs matched (e.g.
comparing spectra from different instruments). Build the index arrays with
dtype=int and add a regression test.

+++ matchms/similarity/MetadataMatch.py
@@ -164,8 +164,8 @@ class MetadataMatch(BaseSimilarity):
                 rows.extend(idx)
                 cols.extend([i] * len(idx))
 
-            rows = np.array(rows)
-            cols = np.array(cols)
+            rows = np.array(rows, dtype=int)
+            cols = np.array(cols, dtype=int)
             scores = np.ones(len(rows))
         else:
             if is_symmetric:
